+++ packages/terra-data-grid/CHANGELOG.md
@@ -13,6 +13,7 @@
   * Added keyboard support for column resizing.
   * Added translations.
   * Added "No Result" cells to FlowsheetDataGrid.
+  * Added ability to select a range of cells to FlowsheetDataGrid.
 
 ## 0.8.1 - (October 17, 2023)